Use unsigned bit position in FaxModule::FaxG4Decode()
Since the only caller for FaxG4Decode() use unsigned bit positions,
just pass that in. Then change FaxModule internals to consistently use
unsigned ints to match.

Reviewed-on: https://pdfium-review.googlesource.com/c/pdfium/+/139812
Commit-Queue: Lei Zhang <thestig@chromium.org>

+++ core/fxcodec/jbig2/JBig2_GrdProc.cpp
@@ -500,7 +500,7 @@ FXCODEC_STATUS CJBig2_GRDProc::StartDecodeMMR(
     progressive_status_ = FXCODEC_STATUS::kError;
     return progressive_status_;
   }
-  int bitpos = static_cast<int>(pStream->getBitPos());
+  uint32_t bitpos = pStream->getBitPos();
   bitpos = FaxModule::FaxG4Decode(pStream->getBufSpan(), bitpos, GBW, GBH,
                                   image->stride(), image_span);
   pStream->setBitPos(bitpos);
